+++ packages/card/src/react/__stories__/index.story.tsx
@@ -1,11 +1,10 @@
-import React from 'react'
-import { storiesOf } from '@storybook/react'
-
 import * as core from '@pluralsight/ps-design-system-core'
 import * as Icon from '@pluralsight/ps-design-system-icon'
-
-import Card, { CardProps } from '..'
 import { StoryFn } from '@storybook/addons'
+import { storiesOf } from '@storybook/react'
+import React from 'react'
+
+import Card, { CardProps } from '../index.js'
 
 const getImgSrc = ({ w = 680, h = 320, id = 42 } = {}) =>
   `//picsum.photos/${w}/${h}?image=${id}&gravity=north`
